[modify] : order_read ui & order:views.py

+++ MOA/order/views.py
@@ -14,7 +14,7 @@ def order_list(request):
         # page = request.GET['page']
     except:
         sort = '0'
-        status = '0'
+        status = '1'
         # page = '1'
 
     # Query String을 통한 정렬
@@ -45,6 +45,9 @@ def order_list(request):
     return render(request, 'order/order_list.html', context)
 
 def order_read(request, pk):
+    if request.user.is_anonymous:
+        return redirect('user:user_login')
+
     order = get_object_or_404(Order, pk=pk)
     fund = None
     if request.user.is_authenticated:
@@ -59,6 +62,9 @@ def order_read(request, pk):
     return render(request, 'order/order_read.html', context)
 
 def order_create(request):
+    if request.user.is_anonymous:
+        return redirect('user:user_login')
+
     if request.method == "POST":
         form = OrderForm(request.POST)
 
@@ -104,8 +110,14 @@ def order_create(request):
 
 
 def order_update(request, pk):
+    if request.user.is_anonymous:
+        return redirect('user:user_login')
+
     order = get_object_or_404(Order, pk=pk)
-    if request.method == "POST":
+    if request.user != order.host:
+        return redirect('order:order_read', order.pk)
+
+    if request.method == "POST" and request.user == order.host:
         form = OrderForm(request.POST, instance=order)
         if form.is_valid():
             order = form.save(commit=False)
@@ -122,16 +134,26 @@ def order_update(request, pk):
 
 
 def order_delete(request, pk):
-    order = get_object_or_404(Order, pk=pk)
+    if request.user.is_anonymous:
+        return redirect('user:user_login')
 
-    if request.method == "POST":
+    order = get_object_or_404(Order, pk=pk)
+    if request.user != order.host:
+        return redirect('order:order_read', order.pk)
+
+    if request.method == "POST" and request.user == order.host:
         order.delete()
         return redirect('order:order_list')
     else:
         redirect('order:order_read', order.pk)
 
 def order_close(request, pk):
+    if request.user.is_anonymous:
+        return redirect('user:user_login')
+
     order = get_object_or_404(Order, pk=pk)
+    if request.user != order.host:
+        return redirect('order:order_read', order.pk)
 
     if request.method == "POST" and request.user == order.host:
